# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Household {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreateHouseholdBody {
  name: string;
}

export interface Storage {
  id: number;
  householdId: number;
  /** @nullable */
  parentId: number | null;
  name: string;
  /** @nullable */
  imageUrl: string | null;
  pathIds: number[];
  pathNames: string[];
  createdBy: string;
  isPublic: boolean;
  deleted: boolean;
  createdAt: string;
}

export interface CreateStorageBody {
  householdId: number;
  /** @nullable */
  parentId?: number | null;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  isPublic?: boolean;
}

export type ItemCustomFields = { [key: string]: unknown };

export interface Item {
  id: number;
  householdId: number;
  storageId: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  category: string | null;
  /** @nullable */
  description: string | null;
  tags: string[];
  customFields: ItemCustomFields;
  locationPathIds: number[];
  locationPathNames: string[];
  createdBy: string;
  isPublic: boolean;
  deleted: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateItemBodyCustomFields = { [key: string]: unknown };

export interface CreateItemBody {
  householdId: number;
  storageId: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  customFields?: CreateItemBodyCustomFields;
  isPublic?: boolean;
}

export interface CategoryCount {
  /** @nullable */
  category: string | null;
  count: number;
}

export interface HouseholdStats {
  totalStorages: number;
  totalItems: number;
  categoryCounts: CategoryCount[];
  recentItems: Item[];
}

export interface UpdateStorageBody {
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  isPublic?: boolean;
}

export type UpdateItemBodyCustomFields = { [key: string]: unknown };

export interface UpdateItemBody {
  storageId?: number;
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  customFields?: UpdateItemBodyCustomFields;
  isPublic?: boolean;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type SearchItemsParams = {
  householdId: number;
  q: string;
};