# packages/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryCount {
  /** @nullable */
  category: string | null;
  count: number;
}