# packages/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateItemBodyCustomFields } from "./createItemBodyCustomFields";

export interface CreateItemBody {
  householdId: number;
  storageId: number;
  name: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  customFields?: CreateItemBodyCustomFields;
  isPublic?: boolean;
}