# packages/api-zod/src/generated/types/householdStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from "./categoryCount";
import type { Item } from "./item";

export interface HouseholdStats {
  totalStorages: number;
  totalItems: number;
  categoryCounts: CategoryCount[];
  recentItems: Item[];
}