# packages/api-zod/src/generated/types/updateItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Home Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateItemBodyCustomFields } from "./updateItemBodyCustomFields";

export interface UpdateItemBody {
  storageId?: number;
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  tags?: string[];
  customFields?: UpdateItemBodyCustomFields;
  isPublic?: boolean;
}